fix: handle errors in flatpak remote delete and ensure vscode repo creation is conditional

+++ home/.chezmoiscripts/linux/run_once_before_01_repos.sh
@@ -11,7 +11,7 @@ sudo dnf install -y \
 # sudo dnf install -y --nogpgcheck --repofrompath 'terra,https://repos.fyralabs.com/terra$releasever' terra-release
 
 # Flathub
-flatpak remote-delete --force fedora
+flatpak remote-delete --force fedora 2>/dev/null || true
 flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
 
 # Copr
@@ -20,7 +20,8 @@ sudo dnf copr enable -y quadratech188/vicinae
 
 # VSCode
 sudo rpm --import https://packages.microsoft.com/keys/microsoft.asc
-sudo tee /etc/yum.repos.d/vscode.repo > /dev/null <<EOF
+if [ ! -f /etc/yum.repos.d/vscode.repo ]; then
+    sudo tee /etc/yum.repos.d/vscode.repo > /dev/null <<EOF
 [code]
 name=Visual Studio Code
 baseurl=https://packages.microsoft.com/yumrepos/vscode
@@ -28,3 +29,4 @@ enabled=1
 gpgcheck=1
 gpgkey=https://packages.microsoft.com/keys/microsoft.asc
 EOF
+fi
